fix update_metadata
The type and value checks must come after we check for None, otherwise they
will fail with the default keyword arguments. Also, we uncomment the reading
and checking of the bloscpack options. Lastly, we need to seek back after
reading the metadata, to make sure it is rewritten into the same place.