hackish fix
Check first if the dtype is a list and then check if the first element is also
a list.  The assumption here is that JSON gives back a list of lists and that
is what needs to be fixed. If we have a list and then a tuple it means the
dtype was probably not serialized and deserialized by JSON. Any dtype that is a
string will not be touched obviously, so we might be able to do away with the
else clause in the _fix functions.

There might still be other corner cases undiscovered.