alternative fix inspired by npy
Instead of serializing the entire dtype struture, i.e. composite python objects
or a simple sting, through JSON, we convert it to a string before handing it
off to JSON. Later, when deserializing we simple call eval on the string and
pass that to numpy. We use ``numpy.safe_eval`` here to ensure only literals are
eval'd and protect us form malicious code.