object dtype for numpy arrays not allowed
The reason is, that for 'O' Numpy arrays, the memory contains pointers to
python objects. Thus, Bloscpack would only serialize those pointers and NOT the
data. This works sometimes for in-memory scenarios, because the Python object
remain allocated and the decompressed pointers point to the correct places.
However this causes a segfault when doing it to disk and keeping arrays over
Python interpreter instances, since the pointers point to memory outside the
current segment (and hence segfault, presumably).