a fix for the recent test failures
As of blosc 1.6.1 this test was failing. Basically, the linspace data
used in the test is no longer compressible by blosc when not using the
shuffle, because the internal threshold has probaby changed. We
circumvent this issue, by using np.zeors which is certainly
compressible, with and without shuffle. The code is now a little bit
irky so more comments were added.